chore: refactors and add support for reading files

+++ token.hpp
@@ -26,8 +26,8 @@ struct Token {
     TokenType type;
     std::string value;
 
-    size_t line;
-    size_t column;
+    int line;
+    int column;
 };
 
 TokenType get_token_type(char c);
